+++ CHANGELOG.md
@@ -1,2 +1,1 @@
-* code theme adjusts to theme
-* web shows error on program termination
+* font size slides for code window